Fix minimal span.
Try e.g. `SuperSmoother().fit(np.arange(80), np.random.rand(80))` (or
anything such that the lowest span (here 0.05*80) is less than 5), which
would previously result in a ValueError.

A minimum span of 5 ensures a span of 3 at the edges (due to the current
handling of the edges) and thus at least 2 points after removing the
center.  Lower spans do not ensure this property and result in fit
failures.

Alternatively, one could also treat edges with non-centered spans
(similarly to `savgol_filter(..., mode="interp")`) -- this is what
Friedman's paper does ("Near the boundaries, it is, of course, not
possible to keep N symmetric." on p.4), but would require a bit more
serious re-architecting of the code.