Fix minimal span.
Try e.g. `SuperSmoother().fit(np.arange(80), np.random.rand(80))` (or
anything such that the lowest span (here 0.05*80) is less than 5), which
would previously result in a ValueError.

At the edges, the span should be non-centered so that the same number of
points are kept.  (This is mentioned in Friedman's paper on page 4: "Near
the boundaries, it is, of course, not possible to keep N symmetric.")
Do so by preventing the indices of the span centers from getting too
close to the edges.

Without this patch, a minimum span of 3 can become a span of only 2
points at the edge; after removing the center (for cross-validation)
only one point is left and fit to a line raises an error.

Non-centered spans are also used elsewhere, e.g. in
`savgol_filter(..., mode="interp")`.